requirements.txt
# Zależności dla systemu analizy sieciowej

# Podstawowe biblioteki
pandas>=1.5.0
numpy>=1.23.0

# Wizualizacje
matplotlib>=3.6.0
seaborn>=0.12.0

# Analiza sieciowa
nfstream>=6.5.0
scapy>=2.5.0

# Machine Learning
scikit-learn>=1.2.0
joblib>=1.2.0

# Threat Intelligence
requests>=2.28.0

# Wizualizacje geograficzne (opcjonalne)
folium>=0.14.0

# CLI
click>=8.1.0

# Reguły Sigma
pyyaml>=6.0
